Added batch in a bag recipes...
Went through and, based on game data, added additional recipes which fit multiple fruit (volume of the plastic bag) per bag.

May look at increasing the time on these recipes but the difference is stuffing several veggies into a bag, sealing it, and irradiating it, same procedure as individual veggy.

Also fixed the original recipes to not give free food from items which have charges.  Requires the same number of charges(items of veggy) of input as the final item output would have.

Adjusted the batch in a bag recipes to fit that fix.

Updated README.md as the batch in a bag recipe is no longer on my todo list.