Balance pass plus vehicle rig.
So now you can make a vehicle rig out of the survivor irradiator (similar to welder) and attach it to vehicles.

Also did a balance pass adding water requirements and tweaking time requirement slightly and battery requirement a lot based on the weight of the veggy/veggies to be irradiated.